requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.0
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
emoji==2.12.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
grpcio==1.65.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
kubernetes==30.1.0
langchain==0.2.11
langchain-anthropic==0.1.21
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langchain-voyageai==0.1.1
langchain-weaviate==0.0.2
langchainhub==0.1.20
langdetect==1.0.9
langsmith==0.1.95
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
marker-pdf==0.2.16
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.37.1
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.2
pdftext==0.3.10
pexpect==4.9.0
pgvector==0.2.5
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.47
psutil==5.9.0
psycopg==3.2.1
psycopg-pool==3.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.3.4
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysqlite3-binary
python-dateutil==2.9.0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich>=10.14.0,<14.0.0
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.1
setuptools==69.5.1
shellingham==1.5.4
simsimd==4.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.2
starlette==0.37.2
streamlit==1.40.1
extra-streamlit-components==0.1.80
surya-ocr==0.4.15
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
texify==0.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
typer==0.12.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.15.0
unstructured-client==0.24.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
validators==0.33.0
voyageai==0.2.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
weaviate-client==4.7.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
